Removes all references to projects in exchange for a more generic "item". Also removes DUAs.
The concept of projects has been moved out of SciAuthZ and into Hypatio. In its place, SciAuthZ now will track permissions that users have on an 'item', which is a generic text field. We will leave it to the external applications (i.e. hypatio) to decide how to name those items.