TC-237: Remove users from SciAuthZ and starts to remove UserPermissionRequest references
Django users are not necessary in SciAuthZ. This commit includes a set of migrations that will move all the User's emails from the UserPermission model to a new "user_email" charfield. Also, this commit starts to remove usages of UserPermissionRequest which do not make sense to store in SciAuthZ. Lastly, this includes an update to the py-auth0-jwt-rest package. The new version of this package will no longer create users if they do not exist. Commented out tests.py which have long now been broken.